Update function to fill n characters

+++ pointers_arrays_strings/0-memset.c
@@ -3,13 +3,14 @@
  * @s: pointer to string
  * @b: character to fill
  * @n: n-th byt to fill
+ * Return: string
  */
 
 char *_memset(char *s, char b, unsigned int n)
 {
-	int i, j;
-	
-	for (j = 0; j < (int) n; j++)
+	unsigned int i, j;
+
+	for (j = 0; j < n; j++)
 	{
 		s[i] = b;
 		i++;
